Agregar seeder cliente, planes,  (Arreglo error de modulo notificacion)

+++ database/migrations/2023_03_28_215504_add_foreign_key_products_table.php
@@ -14,7 +14,7 @@ class AddForeignKeyProductsTable extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->integer('customer_loyalties_id')->unsigned();
+            $table->bigInteger('customer_loyalties_id')->unsigned();
             $table->foreign('customer_loyalties_id')
                  ->references('id')->on('customer_loyalties')->onDelete('cascade') ;
         });
